+++ integration_tests/suite/test_adhoc_conferences.py
@@ -9,6 +9,8 @@ from hamcrest import (
     assert_that,
     calling,
     has_entries,
+    has_item,
+    has_items,
     has_length,
     has_properties,
 )
@@ -42,14 +44,24 @@ class TestAdhocConference(RealAsteriskIntegrationTest):
         self.auth.set_token(MockUserToken(token_id, tenant_uuid=tenant_uuid, user_uuid=user_uuid))
         return token_id
 
-    def given_adhoc_conference(self, user_uuid, participant_count):
+    def given_adhoc_conference(self, *user_uuids, participant_count):
         participant_call_ids = []
+        user_uuids = list(user_uuids)
 
-        host_call_id, participant_call_id = self.real_asterisk.given_bridged_call_stasis(caller_uuid=user_uuid)
+        host_uuid = user_uuids.pop(0)
+        try:
+            participant_uuid = user_uuids.pop(0)
+        except IndexError:
+            participant_uuid = None
+        host_call_id, participant_call_id = self.real_asterisk.given_bridged_call_stasis(caller_uuid=host_uuid, callee_uuid=participant_uuid)
         participant_call_ids.append(participant_call_id)
 
         for _ in range(participant_count - 2):
-            _, participant_call_id = self.real_asterisk.given_bridged_call_stasis(caller_uuid=user_uuid)
+            try:
+                participant_uuid = user_uuids.pop(0)
+            except IndexError:
+                participant_uuid = None
+            _, participant_call_id = self.real_asterisk.given_bridged_call_stasis(caller_uuid=host_uuid, callee_uuid=participant_uuid)
             participant_call_ids.append(participant_call_id)
 
         adhoc_conference = self.calld_client.adhoc_conferences.create_from_user(
@@ -117,12 +129,15 @@ class TestAdhocConference(RealAsteriskIntegrationTest):
                     })))
 
     def test_user_create_adhoc_conference_correct(self):
-        user_uuid = make_user_uuid()
-        token = self.make_user_token(user_uuid)
+        host_uuid = make_user_uuid()
+        participant1_uuid = make_user_uuid()
+        participant2_uuid = make_user_uuid()
+        token = self.make_user_token(host_uuid)
         self.calld_client.set_token(token)
+        host_events = self.bus.accumulator('adhoc_conferences.users.{}.#'.format(host_uuid))
 
-        host_call1_id, participant1_call_id = self.real_asterisk.given_bridged_call_stasis(caller_uuid=user_uuid)
-        host_call2_id, participant2_call_id = self.real_asterisk.given_bridged_call_stasis(caller_uuid=user_uuid)
+        host_call1_id, participant1_call_id = self.real_asterisk.given_bridged_call_stasis(caller_uuid=host_uuid, callee_uuid=participant1_uuid)
+        host_call2_id, participant2_call_id = self.real_asterisk.given_bridged_call_stasis(caller_uuid=host_uuid, callee_uuid=participant2_uuid)
 
         adhoc_conference = self.calld_client.adhoc_conferences.create_from_user(
             host_call1_id,
@@ -144,7 +159,37 @@ class TestAdhocConference(RealAsteriskIntegrationTest):
             }))
             assert_that(host_call2_id, self.c.is_hungup())
         until.assert_(calls_are_bridged, timeout=10)
-        # todo: expect bus event
+
+        def bus_events_are_sent():
+            assert_that(host_events.accumulate(),
+                        has_item(has_entries({
+                            'name': 'adhoc_conference_created',
+                            'data': {
+                                'conference_id': adhoc_conference['conference_id'],
+                            }})))
+            assert_that(host_events.accumulate(), has_items(
+                has_entries({
+                    'name': 'adhoc_conference_participant_joined',
+                    'data': has_entries({
+                        'conference_id': adhoc_conference['conference_id'],
+                        'participant_call': has_entries({
+                            'call_id': participant1_call_id,
+                            'user_uuid': participant1_uuid,
+                        }),
+                    })
+                }),
+                has_entries({
+                    'name': 'adhoc_conference_participant_joined',
+                    'data': has_entries({
+                        'conference_id': adhoc_conference['conference_id'],
+                        'participant_call': has_entries({
+                            'call_id': participant2_call_id,
+                            'user_uuid': participant2_uuid,
+                        }),
+                    })
+                }),
+            ))
+        until.assert_(bus_events_are_sent, timeout=10)
 
     def test_user_create_adhoc_conference_participant_in_conference(self):
         pass
@@ -165,11 +210,15 @@ class TestAdhocConference(RealAsteriskIntegrationTest):
         pass
 
     def test_extra_participant_hangup(self):
-        user_uuid = make_user_uuid()
-        token = self.make_user_token(user_uuid)
+        host_uuid = make_user_uuid()
+        participant1_uuid = make_user_uuid()
+        participant2_uuid = make_user_uuid()
+        token = self.make_user_token(host_uuid)
         self.calld_client.set_token(token)
-        _, call_ids = self.given_adhoc_conference(user_uuid, participant_count=3)
+        adhoc_conference_id, call_ids = self.given_adhoc_conference(host_uuid, participant1_uuid, participant2_uuid, participant_count=3)
         host_call_id, participant1_call_id, participant2_call_id = call_ids
+        host_events = self.bus.accumulator('adhoc_conferences.users.{}.#'.format(host_uuid))
+        participant1_events = self.bus.accumulator('adhoc_conferences.users.{}.#'.format(participant1_uuid))
 
         self.ari.channels.hangup(channelId=participant2_call_id)
 
@@ -183,14 +232,36 @@ class TestAdhocConference(RealAsteriskIntegrationTest):
             assert_that(participant2_call_id, self.c.is_hungup())
         until.assert_(calls_are_still_bridged, timeout=10)
 
-        # todo: expect bus event
+        def bus_events_are_sent():
+            assert_that(host_events.accumulate(),
+                        has_item(has_entries({
+                            'name': 'adhoc_conference_participant_left',
+                            'data': has_entries({
+                                'conference_id': adhoc_conference_id,
+                                'participant_call': has_entries({
+                                    'call_id': participant2_call_id,
+                                    'user_uuid': participant2_uuid,
+                                }),
+                            })})))
+            assert_that(participant1_events.accumulate(),
+                        has_item(has_entries({
+                            'name': 'adhoc_conference_participant_left',
+                            'data': has_entries({
+                                'conference_id': adhoc_conference_id,
+                                'participant_call': has_entries({
+                                    'call_id': participant2_call_id,
+                                    'user_uuid': participant2_uuid,
+                                }),
+                            })})))
+        until.assert_(bus_events_are_sent, timeout=10)
 
     def test_last_participant_hangup(self):
         user_uuid = make_user_uuid()
         token = self.make_user_token(user_uuid)
         self.calld_client.set_token(token)
-        _, call_ids = self.given_adhoc_conference(user_uuid, participant_count=2)
+        adhoc_conference_id, call_ids = self.given_adhoc_conference(user_uuid, participant_count=2)
         host_call_id, participant1_call_id = call_ids
+        host_events = self.bus.accumulator('adhoc_conferences.users.{}.#'.format(user_uuid))
 
         self.ari.channels.hangup(channelId=participant1_call_id)
 
@@ -199,12 +270,22 @@ class TestAdhocConference(RealAsteriskIntegrationTest):
             assert_that(participant1_call_id, self.c.is_hungup())
         until.assert_(calls_are_hungup, timeout=10)
 
+        def bus_events_are_sent():
+            assert_that(host_events.accumulate(),
+                        has_item(has_entries({
+                            'name': 'adhoc_conference_deleted',
+                            'data': {
+                                'conference_id': adhoc_conference_id,
+                            }})))
+        until.assert_(bus_events_are_sent, timeout=10)
+
     def test_host_hangup(self):
         user_uuid = make_user_uuid()
         token = self.make_user_token(user_uuid)
         self.calld_client.set_token(token)
-        _, call_ids = self.given_adhoc_conference(user_uuid, participant_count=3)
+        adhoc_conference_id, call_ids = self.given_adhoc_conference(user_uuid, participant_count=3)
         host_call_id, participant1_call_id, participant2_call_id = call_ids
+        host_events = self.bus.accumulator('adhoc_conferences.users.{}.#'.format(user_uuid))
 
         self.ari.channels.hangup(channelId=host_call_id)
 
@@ -213,3 +294,12 @@ class TestAdhocConference(RealAsteriskIntegrationTest):
             assert_that(participant1_call_id, self.c.is_hungup())
             assert_that(participant2_call_id, self.c.is_hungup())
         until.assert_(calls_are_hungup, timeout=10)
+
+        def bus_events_are_sent():
+            assert_that(host_events.accumulate(),
+                        has_item(has_entries({
+                            'name': 'adhoc_conference_deleted',
+                            'data': {
+                                'conference_id': adhoc_conference_id,
+                            }})))
+        until.assert_(bus_events_are_sent, timeout=10)