+++ wazo_calld/plugins/adhoc_conferences/stasis.py
@@ -4,41 +4,46 @@
 import threading
 
 from ari.exceptions import ARINotFound
-from wazo_calld.ari_ import DEFAULT_APPLICATION_NAME
-from wazo_calld.plugin_helpers.ari_ import Bridge
+from wazo_calld.plugin_helpers.ari_ import Bridge, Channel
+from wazo_calld.plugins.calls.services import CallsService
 
 import logging
 
 logger = logging.getLogger(__name__)
 
+ADHOC_CONFERENCE_STASIS_APP = 'adhoc_conference'
+
 
 class AdhocConferencesStasis:
 
-    def __init__(self, ari):
+    def __init__(self, ari, notifier):
         self.ari = ari.client
         self._core_ari = ari
+        self._notifier = notifier
         self.adhoc_conference_creation_lock = threading.Lock()
 
     def _subscribe(self):
         self.ari.on_channel_event('StasisStart', self.on_stasis_start)
+        self.ari.on_channel_event('ChannelEnteredBridge', self.on_channel_entered_bridge)
         self.ari.on_channel_event('ChannelLeftBridge', self.on_channel_left_bridge)
+        self.ari.on_bridge_event('BridgeDestroyed', self.on_bridge_destroyed)
 
     def initialize(self):
         self._subscribe()
-        self._core_ari.register_application(DEFAULT_APPLICATION_NAME)
+        self._core_ari.register_application(ADHOC_CONFERENCE_STASIS_APP)
 
     def on_stasis_start(self, event_objects, event):
+        if event['application'] != ADHOC_CONFERENCE_STASIS_APP:
+            return
+
         logger.debug('on_stasis_start: %(id)s (%(name)s)', event['channel'])
         try:
-            sub_app, adhoc_conference_id = event['args']
+            (adhoc_conference_id,) = event['args']
         except ValueError:
             logger.debug('ignoring StasisStart event: channel %s, app %s, args %s',
                          event['channel']['name'],
                          event['application'],
                          event['args'])
-            return
-
-        if sub_app != 'adhoc_conference':
             return
 
         with self.adhoc_conference_creation_lock:
@@ -50,39 +55,85 @@ class AdhocConferencesStasis:
                     type='mixing',
                     bridgeId=adhoc_conference_id,
                 )
+                self.ari.applications.subscribe(
+                    applicationName=ADHOC_CONFERENCE_STASIS_APP,
+                    eventSource=f'bridge:{bridge.id}',
+                )
+
         channel_id = event['channel']['id']
         logger.debug('adhoc conference %s: bridging participant %s', adhoc_conference_id, channel_id)
         bridge.addChannel(channel=channel_id)
 
-        channel = event_objects['channel']
+    def on_channel_entered_bridge(self, channel, event):
+        if event['application'] != ADHOC_CONFERENCE_STASIS_APP:
+            return
+
+        adhoc_conference_id = event['bridge']['id']
+        channel_id = event['channel']['id']
+
+        logger.debug('adhoc conference %s: channel %s entered', adhoc_conference_id, channel_id)
+
         try:
             is_adhoc_conference_host = channel.getChannelVar(variable='WAZO_IS_ADHOC_CONFERENCE_HOST')['value'] == 'true'
         except ARINotFound:
             logger.error('adhoc conference %s: channel %s hungup too early or variable not found', adhoc_conference_id, channel_id)
             return
 
+        bridge_helper = Bridge(adhoc_conference_id, self.ari)
+        participant_call = CallsService.make_call_from_channel(self. ari, channel)
+        other_participant_uuids = bridge_helper.valid_user_uuids()
+        self._notifier.participant_joined(adhoc_conference_id, other_participant_uuids, participant_call)
+
         if is_adhoc_conference_host:
-            Bridge(adhoc_conference_id, self.ari).global_variables.set('WAZO_ADHOC_CONFERENCE_HOST', channel_id)
+            bridge_helper.global_variables.set('WAZO_HOST_CHANNEL_ID', channel_id)
+            host_user_uuid = Channel(channel_id, self.ari).user()
+            bridge_helper.global_variables.set('WAZO_HOST_USER_UUID', host_user_uuid)
+
+            self._notify_host_of_channels_already_present(adhoc_conference_id, event['bridge']['channels'], host_user_uuid)
+
+    def _notify_host_of_channels_already_present(self, adhoc_conference_id, channel_ids, host_user_uuid):
+        for channel_id in channel_ids:
+            try:
+                other_participant_channel = self.ari.channels.get(channelId=channel_id)
+            except ARINotFound:
+                logger.error('adhoc conference %s: participant %s hanged up before host arrived', adhoc_conference_id, channel_id)
+                continue
+            other_participant_call = CallsService.make_call_from_channel(self.ari, other_participant_channel)
+            self._notifier.participant_joined(adhoc_conference_id, [host_user_uuid], other_participant_call)
 
     def on_channel_left_bridge(self, channel, event):
+        if event['application'] != ADHOC_CONFERENCE_STASIS_APP:
+            return
+
         adhoc_conference_id = event['bridge']['id']
         channel_id = event['channel']['id']
 
-        if not self._channel_left_adhoc_conference(channel_id, adhoc_conference_id):
-            logger.debug('adhoc conference: ignoring channel %s left bridge', channel_id)
-            return
-
         logger.debug('adhoc conference %s: channel %s left', adhoc_conference_id, channel_id)
 
-        try:
-            adhoc_conference_host_channel_id = Bridge(adhoc_conference_id, self.ari).global_variables.get(variable='WAZO_ADHOC_CONFERENCE_HOST')
-        except KeyError:
-            logger.error('adhoc conference %s: could not find conference')
-            return
+        bridge_helper = Bridge(adhoc_conference_id, self.ari)
 
-        if adhoc_conference_host_channel_id == channel_id:
-            logger.debug('adhoc conference %s: host %s left, hanging up all participants', adhoc_conference_id, channel_id)
-            self._hangup_all_participants(adhoc_conference_id)
+        participant_call = CallsService.make_call_from_dead_channel(channel)
+        other_participant_uuids = bridge_helper.valid_user_uuids()
+        self._notifier.participant_left(adhoc_conference_id, other_participant_uuids, participant_call)
+
+        if bridge_helper.has_lone_channel():
+            logger.debug('adhoc conference %s: only one participant %s left, hanging up', adhoc_conference_id, channel_id)
+            bridge_helper.hangup_all()
+        elif bridge_helper.is_empty():
+            logger.debug('adhoc conference %s: bridge is empty, destroying', adhoc_conference_id)
+            try:
+                self.ari.bridges.destroy(bridgeId=adhoc_conference_id)
+            except ARINotFound:
+                pass
+        else:
+            try:
+                adhoc_conference_host_channel_id = bridge_helper.global_variables.get(variable='WAZO_HOST_CHANNEL_ID')
+            except KeyError:
+                logger.error('adhoc conference %s: could not retrieve host channel id', adhoc_conference_id)
+                return
+            if adhoc_conference_host_channel_id == channel_id:
+                logger.debug('adhoc conference %s: host %s left, hanging up all participants', adhoc_conference_id, channel_id)
+                self._hangup_all_participants(adhoc_conference_id)
 
     def _hangup_all_participants(self, adhoc_conference_id):
         try:
@@ -99,15 +150,20 @@ class AdhocConferencesStasis:
             except ARINotFound:
                 pass
 
-        logger.debug('adhoc conference %s: destroying bridge', adhoc_conference_id)
-        try:
-            self.ari.bridges.destroy(bridgeId=adhoc_conference_id)
-        except ARINotFound:
-            pass
+    def on_bridge_destroyed(self, bridge, event):
+        if event['application'] != ADHOC_CONFERENCE_STASIS_APP:
+            return
 
-    def _channel_left_adhoc_conference(self, channel_id, bridge_id):
+        logger.debug('adhoc conference %s: bridge was destroyed', bridge.id)
+
         try:
-            Bridge(bridge_id, self.ari).global_variables.get(variable='WAZO_ADHOC_CONFERENCE_HOST')
+            host_user_uuid = Bridge(bridge.id, self.ari).global_variables.get('WAZO_HOST_USER_UUID')
         except KeyError:
-            return False
-        return True
+            logger.error('adhoc conference %s: could not retrieve host user uuid', bridge.id)
+            return
+
+        self._notifier.deleted(bridge.id, host_user_uuid)
+
+        bridge_helper = Bridge(bridge.id, self.ari)
+        bridge_helper.global_variables.unset('WAZO_HOST_USER_UUID')
+        bridge_helper.global_variables.unset('WAZO_HOST_CHANNEL_ID')