+++ CHANGELOG.md
@@ -5,6 +5,13 @@
 * New API to create adhoc conferences
 
   * `POST /users/me/conferences/adhoc`
+  
+* New events for adhoc conferences:
+
+  * `adhoc_conference_created`
+  * `adhoc_conference_deleted`
+  * `adhoc_conference_participant_joined`
+  * `adhoc_conference_participant_left`
 
 * New API to answer calls